+++ api/app/controllers/spree/api/addresses_controller.rb
@@ -13,31 +13,35 @@ module Spree
         authorize! :update, @order, order_token
         find_address
 
-        if @address.update_attributes(address_params)
+        if @order.update_attributes({"#{@order_source}_attributes" => address_params})
+          @address = @order.send(@order_source)
           respond_with(@address, :default_template => :show)
         else
+          @address = @order.send(@order_source)
           invalid_resource!(@address)
         end
       end
 
       private
-        def address_params
-          params.require(:address).permit(permitted_address_attributes)
-        end
+      def address_params
+        params.require(:address).permit(permitted_address_attributes)
+      end
 
-        def find_order
-          @order = Spree::Order.find_by!(number: order_id)
-        end
+      def find_order
+        @order = Spree::Order.find_by!(number: order_id)
+      end
 
-        def find_address
-          @address = if @order.bill_address_id == params[:id].to_i
-            @order.bill_address
-          elsif @order.ship_address_id == params[:id].to_i
-            @order.ship_address
-          else
-            raise CanCan::AccessDenied
-          end
+      def find_address
+        @address = if @order.bill_address_id == params[:id].to_i
+          @order_source = :bill_address
+          @order.bill_address
+        elsif @order.ship_address_id == params[:id].to_i
+          @order_source = :ship_address
+          @order.ship_address
+        else
+          raise CanCan::AccessDenied
         end
+      end
     end
   end
 end