+++ backend/app/controllers/spree/admin/orders/customer_details_controller.rb
@@ -9,7 +9,7 @@ module Spree
         end
 
         def edit
-          country_id = Address.default.country.id
+          country_id = Country.default.id
           @order.build_bill_address(:country_id => country_id) if @order.bill_address.nil?
           @order.build_ship_address(:country_id => country_id) if @order.ship_address.nil?