+++ core/app/models/spree/user_address.rb
@@ -1,0 +1,9 @@
+module Spree
+  class UserAddress < ActiveRecord::Base
+    belongs_to :user, class_name: Spree.user_class.to_s, foreign_key: "user_id"
+    belongs_to :address, class_name: "Spree::Address"
+
+    validates_uniqueness_of :address_id, scope: :user_id
+    validates_uniqueness_of :user_id, conditions: -> { where(default: true) }, message: :default_address_exists, if: :default?
+  end
+end